+++ frontend/CHANGELOG.md
@@ -7,6 +7,8 @@ and this project adheres to [Semantic Versioning](http://semver.org/).
 The latest version of this document is always available in
 [releases][releases-url].
 
+## [next]
+- fix: Fixed bug in history view preventing toggle-view.
 
 ## [3.2.0]
 - feat: Initial beta support for variants